+++ src/ibex_imaging_knowledge_base_utilities/argparse_types.py
@@ -18,6 +18,7 @@
 
 import pathlib
 import argparse
+import pandas as pd
 
 # definitions of argparse types, enables argparse to validate the command line parameters
 
@@ -31,6 +32,33 @@ def file_path_endswith(path_str, ends_with_pattern=""):
             f"Invalid argument ({path_str}), not a file path, file does not exist, or path does not"
             + f' end with "{ends_with_pattern}".'
         )
+
+
+def csv_path(path, required_columns={}):
+    """
+    Define the csv_path type for use with argparse. Checks
+    that the given path string is a path to a csv file and that the
+    header of the csv file contains the required columns.
+    """
+    p = pathlib.Path(path)
+    required_columns = set(required_columns)
+    if p.is_file():
+        try:  # only read the csv header
+            expected_columns_exist = required_columns.issubset(
+                set(pd.read_csv(path, nrows=0).columns.tolist())
+            )
+            if expected_columns_exist:
+                return p
+            else:
+                raise argparse.ArgumentTypeError(
+                    f"Invalid argument ({path}), does not contain all expected columns."
+                )
+        except UnicodeDecodeError:
+            raise argparse.ArgumentTypeError(
+                f"Invalid argument ({path}), not a csv file."
+            )
+    else:
+        raise argparse.ArgumentTypeError(f"Invalid argument ({path}), not a file.")
 
 
 def dir_path(path):